+++ AtlasLootClassic/Button/Item_type.lua
@@ -43,11 +43,13 @@ ClickHandler:Add(
 		DressUp = { "LeftButton", "Ctrl" },
 		SetFavourite = { "LeftButton", "Alt" },
 		ShowExtraItems = { "LeftButton", "None" },
+		WoWHeadLink = { "RightButton", "Shift" },
 		types = {
 			ChatLink = true,
 			DressUp = true,
 			ShowExtraItems = true,
 			SetFavourite = true,
+			WoWHeadLink = true,
 		},
 	},
 	{
@@ -55,6 +57,7 @@ ClickHandler:Add(
 		{ "DressUp", 		AL["Dress up"], 			AL["Shows the item in the Dressing room"] },
 		{ "SetFavourite", 	AL["Set Favourite"], 		AL["Set the item as favourite"] },
 		{ "ShowExtraItems", AL["Show extra items"], 	AL["Shows extra items (tokens,mats)"] },
+		{ "WoWHeadLink", 	AL["Show WowHead link"], 	AL["Shows a copyable link for WoWHead"] },
 	}
 )
 
@@ -117,6 +120,8 @@ function Item.OnMouseAction(button, mouseButton)
 		local itemInfo, itemLink = GetItemInfo(button.ItemString or button.ItemID)
 		itemLink = itemLink or button.ItemString
 		AtlasLoot.Button:AddChatLink(itemLink or "item:"..button.ItemID)
+	elseif mouseButton == "WoWHeadLink" then
+		AtlasLoot.Button:OpenWoWHeadLink(button, "item", button.ItemID)
 	elseif mouseButton == "DressUp" then
 		local itemInfo, itemLink = GetItemInfo(button.ItemString or button.ItemID)
 		itemLink = itemLink or button.ItemString