+++ AtlasLootClassic/ItemDB/ItemDB.lua
@@ -411,6 +411,12 @@ end
 -- ##################################################
 --	ContentProto
 -- ##################################################
+local ATLAS_TEXTURE, PATH_TEXTURE = "|A:%s:0:0|a ","|T%s:0|t "
+local SpecialMobList = {
+	rare = format(ATLAS_TEXTURE, "nameplates-icon-elite-silver"),
+	elite = format(ATLAS_TEXTURE, "nameplates-icon-elite-gold"),
+}
+
 --- Get the content Type
 -- @return ContentName, ContentIndex
 function ItemDB.ContentProto:GetContentType()
@@ -426,25 +432,25 @@ function ItemDB.ContentProto:GetName(raw)
 	if self.AreaID and not self.MapID then
 		self.MapID = self.AreaID
 	end
-	local add = ""
+	local addEnd = ""
 	if not raw then
 		if AtlasLoot.db.showLvlRange and self.LevelRange then
 			if AtlasLoot.db.showMinEnterLvl then
-				add = format(LEVEL_RANGE_FORMAT, self.LevelRange[1] or 0, self.LevelRange[2] or 0, self.LevelRange[3] or 0 )
+				addEnd = format(LEVEL_RANGE_FORMAT, self.LevelRange[1] or 0, self.LevelRange[2] or 0, self.LevelRange[3] or 0 )
 			else
-				add = format(LEVEL_RANGE_FORMAT2, self.LevelRange[2] or 0, self.LevelRange[3] or 0 )
+				addEnd = format(LEVEL_RANGE_FORMAT2, self.LevelRange[2] or 0, self.LevelRange[3] or 0 )
 			end
 		end
 		if AtlasLoot.db.ContentPhase.enableOnLootTable and self.ContentPhase and not ContentPhase:IsActive(self.ContentPhase) then
-			add = add.."  "..format(CONTENT_PHASE_FORMAT, self.ContentPhase)
+			addEnd = addEnd.."  "..format(CONTENT_PHASE_FORMAT, self.ContentPhase)
 		end
 	end
 	if self.name then
-		return self.name..add
+		return self.name..addEnd
 	elseif self.MapID then
-		return C_Map.GetAreaInfo(self.MapID)..add or "MapID:"..self.MapID
+		return C_Map.GetAreaInfo(self.MapID)..addEnd or "MapID:"..self.MapID
 	elseif self.FactionID then
-		return AtlasLoot:Faction_GetFactionName(self.FactionID)..add
+		return AtlasLoot:Faction_GetFactionName(self.FactionID)..addEnd
 	else
 		return UNKNOWN
 	end
@@ -463,16 +469,19 @@ function ItemDB.ContentProto:GetNameForItemTable(index, raw)
 	if raw and not self.items[index] then return end
 	assert(index and self.items[index], "index not found.")
 	index = self.items[index]
-	local add = ""
+	local addStart, addEnd = "", ""
 	if not raw then
 		if AtlasLoot.db.ContentPhase.enableOnLootTable and index.ContentPhase and not ContentPhase:IsActive(index.ContentPhase) then
-			add = add.." "..format(CONTENT_PHASE_FORMAT, index.ContentPhase)
+			addEnd = addEnd.." "..format(CONTENT_PHASE_FORMAT, index.ContentPhase)
+		end
+		if index.specialType and SpecialMobList[index.specialType] then
+			addStart = SpecialMobList[index.specialType]
 		end
 	end
 	if index.name then
-		return index.name..add
+		return addStart..index.name..addEnd
 	elseif index.FactionID then
-		return GetFactionInfoByID(index.FactionID)..add --or "Faction "..self.items[index].FactionID
+		return addStart..GetFactionInfoByID(index.FactionID)..addEnd --or "Faction "..self.items[index].FactionID
 	else
 		return UNKNOWN
 	end