+++ AtlasLootClassic/Documentation/Release_Notes.md
@@ -2,6 +2,11 @@
 
 This mod is distributed under Version 2 of the GPL.  A copy of the GPL is included in this zip file with links to non-english translations.
 
+## v1.1.7-beta (Aug. 27, 2019)
+
+- Translation update from curseforge
+- Some small fixes
+
 ## v1.1.6-beta (Aug. 25, 2019)
 
 - Big thanks for the translation work at curseforge!