+++ CHANGELOG.md
@@ -4,7 +4,9 @@ This mod is distributed under Version 2 of the GPL.  A copy of the GPL is includ
 
 [Changelog history](https://github.com/Hoizame/AtlasLootClassic/blob/master/AtlasLootClassic/Documentation/Release_Notes.md)
 
-## v1.1.7-beta (Aug. 27, 2019)
+## v1.1.8-beta (Aug. 31, 2019)
 
-- Translation update from curseforge
-- Some small fixes
+- Split Scarlet Monastery into wings
+- Add Rare indicator for some bosses
+- Add support for WoW-Head links ( Shift + Right Click )
+- Some content phase fixes